+++ docs/CHANGELOG.md
@@ -10,6 +10,9 @@ ones are marked like "v1.0.0-fork".
 ### Added
 
 * `unloadformcheck.js` no declares a new object `lwt_form_check` that contains all the functions needed.
+* New globals ([#163](https://github.com/HugoFara/lwt/issues/163)):
+  * On `inc/kernel_utility.php`: `LWT_APP_VERSION` and `LWT_RELEASE_DATE`.
+  * On `api.php`: `LWT_API_VERSION` and `LWT_API_RELEASE_DATE`.
 
 ### Fixed